Seperated Init
Updated pathways, must be tested
